Add files via upload
The test-bench codes written in Hardware Description Language (HDL) are run in the Model-sim-
Altera software to evaluate the behaviour of Design Under Test (DUT). In this process, four of the
components are simulated to check the behaviour with their test benches, which are derived from
there truth table. The four components are listed below with their HDL code, test-bench code,
simulation waveform, and the transcripts for matching the output.